renamed FRXdrive to FRXDrive moved config to opmode/config
