+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/FRXDriveTest.java
@@ -1,18 +1,15 @@
 package org.firstinspires.ftc.teamcode.opmode;
 
-import com.acmerobotics.dashboard.DashboardCore;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.teamcode.config.XDriveConfig;
+import org.firstinspires.ftc.teamcode.opmode.config.XDriveConfig;
 import org.firstinspires.ftc.teamcode.modules.drive.HDrive;
 import org.firstinspires.ftc.teamcode.modules.drive.XDrive;
 @TeleOp
-public class FRXdriveTest extends LinearOpMode {
+public class FRXDriveTest extends LinearOpMode {
     public void runOpMode() {
         telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
         XDriveConfig xDriveConfig = new XDriveConfig(hardwareMap);
