+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/config/ServoMotorTestConfig.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.config;
+package org.firstinspires.ftc.teamcode.opmode.config;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
